Refactored updateReport command
